refactoring renderre
