+++ cursor/misc.py
@@ -1,6 +1,3 @@
-import cursor.data as data
-import cursor.renderer
-
 import time
 import numpy as np
 import inspect
@@ -111,28 +108,6 @@ def map(value, inputMin, inputMax, outputMin, outputMax, clamp):
         elif outVal < outputMin:
             outVal = outputMin
     return outVal
-
-
-def save_wrapper(pc, projname, fname):
-    jpeg_folder = data.DataDirHandler().jpg(projname)
-    jpeg_renderer = cursor.renderer.JpegRenderer(jpeg_folder)
-
-    jpeg_renderer.render(pc, scale=1.0)
-    jpeg_renderer.save(fname)
-
-    svg_folder = data.DataDirHandler().svg(projname)
-    svg_renderer = cursor.renderer.SvgRenderer(svg_folder)
-
-    svg_renderer.render(pc)
-    svg_renderer.save(fname)
-
-
-def save_wrapper_jpeg(pc, projname, fname, scale=4.0, thickness=3):
-    folder = data.DataDirHandler().jpg(projname)
-    jpeg_renderer = cursor.renderer.JpegRenderer(folder)
-
-    jpeg_renderer.render(pc, scale=scale, thickness=thickness)
-    jpeg_renderer.save(fname)
 
 
 def current_source(frame):
